arreglo de 4 detalles de nombres incorrectos a lo que pide la practica

+++ Backend/daw2026/src/main/java/daw2026/Dto/CreateEventRequest.java
@@ -1,5 +1,6 @@
 package daw2026.Dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import daw2026.Model.Category;
@@ -17,7 +18,7 @@ public class CreateEventRequest {
     private Category category;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private String address;
 }
